+++ src/main/kotlin/com/nylas/models/ContactGroupId.kt
@@ -1,0 +1,8 @@
+package com.nylas.models
+
+/**
+ * Class representing an object that points to a contact group ID.
+ */
+data class ContactGroupId(
+  val id: String,
+)