+++ src/main/kotlin/com/nylas/models/GroupType.kt
@@ -1,0 +1,14 @@
+package com.nylas.models
+
+import com.squareup.moshi.Json
+
+enum class GroupType {
+  @Json(name = "user")
+  USER,
+
+  @Json(name = "system")
+  SYSTEM,
+
+  @Json(name = "other")
+  OTHER,
+}